+++ inst/template/R/_disable_autoload.R
@@ -1,0 +1,1 @@
+# See ?shiny::loadSupport